requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
astroid==2.8.4
async-timeout==4.0.0
attrs==21.2.0
charset-normalizer==2.0.7
frozenlist==1.2.0
graia-ariadne==0.3.9
graia-broadcast==0.13.1
idna==3.3
isort==5.10.1
lazy-object-proxy==1.6.0
loguru==0.5.3
mccabe==0.6.1
multidict==5.2.0
platformdirs==2.4.0
pydantic==1.8.2
pylint==2.11.1
toml==0.10.2
typing-extensions==3.10.0.2
wrapt==1.13.3
yarl==1.7.2
